experimental cuda backend
needs substantial work.  Should probably wrap GPU arrays with a technique that can keep them in gpu memory but allow a numpy copy in RAM to be made.  This way e.g. fftshift can be done on the GPU, and most of the compute time is already spent shuffling memory; the GPU has far greater bandwidth to its own registers.