MTF vs Field plots, baby phase retrieval
baby phase retrieval uses MTF data taken on the optical axis to estimate W020, W040, W060, and W080 terms.  Currently the balance of these terms is unconstrained, so e.g. the condition of best focus is not guaranteed.  May port autofocus function to the Pupil class to allow autofocus during optimization.  Preliminary results suggest this is not a large issue.